+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add new operations, based on Sekoia API 3.1 documentation and automation-library playbook action JSON files:
     - `edit_case` operation
     - `get_case` operation
+    - `get_custom_priority` operation
     - `list_assets` operation
     - `revoke_assetv2` operation
     - `search_cases` operation